refactor(direct-control): prune stale facade fs import
Refs: openspec/changes/civ7-support-direct-control-modularization

Removes the stale node:fs/promises import from the public direct-control facade after proof/log and capability catalog modules took ownership of filesystem-backed reads. Public facade exports remain stable and runtime behavior is unchanged.

This keeps procedure schemas, operation/proof telemetry, AI ingestion, semantic CLI projection, hotseat runtime proof, Effect/Bun composition, Effect/oRPC procedure cores, and Task 2.9.4 matrix-row acceptance pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- rg ownership scan for stale filesystem imports in packages/civ7-direct-control/src/index.ts
- git diff --check
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts public-api.test.ts session.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --cached --check

Local source hygiene proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, Effect/oRPC, or procedure-core proof claimed.
